requirements-cuda.txt
numpy
pandas
tensorboard
pytorch-lightning
matplotlib
torchmetrics
datasets